Some small changes
