+++ Projects/CoX/Servers/MapServer/Events/MapXferWait.h
@@ -1,0 +1,38 @@
+/*
+ * SEGS - Super Entity Game Server
+ * http://www.segs.io/
+ * Copyright (c) 2006 - 2018 SEGS Team (see AUTHORS.md)
+ * This software is licensed under the terms of the 3-clause BSD License. See LICENSE.md for details.
+ */
+
+#pragma once
+#include "GameCommand.h"
+#include "MapEventTypes.h"
+
+#include <QtCore/QString>
+
+namespace SEGSEvents
+{
+    // [[ev_def:type]]
+class MapXferWait final : public MapLinkEvent
+{
+public:
+                MapXferWait() : MapLinkEvent(MapEventTypes::evMapXferWait)
+                {}
+                MapXferWait(QString map_name) : MapLinkEvent(MapEventTypes::evMapXferWait)
+                {
+                    m_map_name = map_name;
+                }
+
+    void        serializeto(BitStream &bs) const override {
+                    bs.StorePackedBits(1, 10);  // opcode 
+                    bs.StoreString(m_map_name);
+                }
+    void        serializefrom(BitStream &src) override
+                {
+                }
+    EVENT_IMPL(MapXferWait)
+            // [[ev_def:field]]
+    QString     m_map_name;
+};
+} //end of SEGSEvents namespace
